+++ src/gwt/src/org/rstudio/studio/client/workbench/views/source/editors/text/ace/AceEditorCommandDispatcher.java
@@ -80,6 +80,14 @@ public class AceEditorCommandDispatcher
             ExecutionPolicy.FOCUSED);
    }
    
+   @Handler
+   public void onInsertPipeOperator()
+   {
+      fireEvent(
+            AceEditorCommandEvent.CommandType.INSERT_PIPE_OPERATOR,
+            ExecutionPolicy.FOCUSED);
+   }
+   
    // Private methods ----
    
    private void fireEvent(CommandType type, ExecutionPolicy policy)