+++ NEWS.md
@@ -29,7 +29,8 @@
 * Expand selection command learns how to expand to current statement
 * Added (as unbound command) Sublime Text style 'Quick Add Next'
 * The various 'yank' commands are now rebindable (Ctrl + Y, Ctrl + K, Ctrl + U)
-* Insert assignment operator (Alt+-) is now rebindable
+* Insert assignment operator (Alt+-, '<-') is now rebindable
+* Insert pipe operator (Cmd+Shift+M, '%>%') is now rebindable
 
 ### R Markdown